move new html image loader to its own script
