+++ crates/notan_glow/src/prelude.rs
@@ -1,1 +1,2 @@
-pub use crate::texture_source::*;
+#[cfg(target_arch = "wasm32")]
+pub use crate::html_image::*;
